Move containers inside /components
